+++ R/McmcOptions-methods.R
@@ -1,51 +1,80 @@
-#####################################################################################
-## Author: Daniel Sabanes Bove [sabanesd *a*t* roche *.* com]
-## Project: Object-oriented implementation of CRM designs
-##
-## Time-stamp: <[McmcOptions-methods.R] by DSB Fre 16/01/2015 11:58>
-##
-## Description:
-## Functions/methods for the MCMC formal class.
-##
-## History:
-## 12/12/2013   file creation (copy from glmBfp package)
-#####################################################################################
+#' @include McmcOptions-class.R
+NULL
 
-##' Compute the number of samples for a given MCMC options triple
-##'
-##' @param mcmcOptions the \code{\linkS4class{McmcOptions}} object
-##' @return the resulting sample size
-##'
-##' @example examples/McmcOptions-methods-sampleSize.R
-##' @export
-##' @keywords programming
-sampleSize <-
-    function(mcmcOptions)
-{
-    stopifnot(is(mcmcOptions, "McmcOptions"))
+# sampleSize ----
 
-    return(as.integer(ceiling((mcmcOptions@iterations - mcmcOptions@burnin) /
-                                  mcmcOptions@step)))
-}
+#' Computing the Number of Samples
+#'
+#' @description `r lifecycle::badge("stable")`
+#'
+#' A method that computes the number of samples for a given MCMC options triple.
+#'
+#' @param object (`McmcOptions`)\cr object based on which the number of samples
+#'   is computed.
+#' @param ... not used.
+#' @return A number of samples for a given MCMC options.
+#' @export
+#'
+setGeneric(
+  name = "sampleSize",
+  def = function(object, ...) {
+    standardGeneric("sampleSize")
+  },
+  valueClass = "integer"
+)
 
-##' Determine if we should save this sample
-##'
-##' @param iteration the current iteration index
-##' @param mcmcOptions  the \code{\linkS4class{McmcOptions}} object
-##' @return Logical value, if we should save this sample
-##'
-##' @example examples/McmcOptions-methods-saveSample.R
-##' @export
-##' @keywords programming internal
-saveSample <-
-    function(iteration,
-             mcmcOptions)
-{
-    stopifnot(is(mcmcOptions, "McmcOptions"))
+# McmcOptions-sampleSize ----
 
-    return((iteration > mcmcOptions@burnin) &&
-           (((iteration - mcmcOptions@burnin) %% mcmcOptions@step) == 0))
-}
+#' @rdname sampleSize
+#' @aliases sampleSize-McmcOptions-method
+#' @example examples/McmcOptions-methods-sampleSize.R
+setMethod(
+  f = "sampleSize",
+  signature = signature(object = "McmcOptions"),
+  definition = function(object, ...) {
+    iterations_relative <- object@iterations - object@burnin
+    if (iterations_relative <= 0) {
+      return(0L)
+    }
+    safeInteger(floor(iterations_relative / object@step))
+  }
+)
 
+# saveSample ----
 
+#' Determining if this Sample Should be Saved
+#'
+#' @description `r lifecycle::badge("stable")`
+#'
+#' A method that determines if a sample from a given `iteration` should be
+#'   saved. The sample should be saved if and only if:
+#'   it is not in burn-in period and it matches the `step`.
+#'
+#' @param object (`McmcOptions`)\cr object based on which the answer is
+#'   determined.
+#' @param iteration (`count`)\cr the current iteration index.
+#' @param ... not used.
+#' @return `TRUE` if this sample should be saved.
+#' @export
+#'
+setGeneric(
+  name = "saveSample",
+  def = function(object, iteration, ...) {
+    standardGeneric("saveSample")
+  },
+  valueClass = "logical"
+)
 
+# McmcOptions-saveSample ----
+
+#' @rdname saveSample
+#' @aliases saveSample-McmcOptions-method
+#' @example examples/McmcOptions-methods-saveSample.R
+setMethod(
+  f = "saveSample",
+  signature = signature(object = "McmcOptions"),
+  definition = function(object, iteration, ...) {
+    iteration_relative <- iteration - object@burnin
+    iteration_relative > 0 && ((iteration_relative %% object@step) == 0)
+  }
+)