+++ R/mcmc.R
@@ -695,7 +695,7 @@ setMethod("mcmc",
 
                 ##4)Save Samples
 
-                if (saveSample(iterMcmc,options)){
+                if (saveSample(options, iterMcmc)){
                   samples$ExpEff[iterSave,]<-betaW
                   samples$sigma2[iterSave,1]<-sigma2
                   samples$sigma2betaW[iterSave,1] <-sigma2betaW